+++ moa/src/main/java/com/korit/moa/moa/controller/MeetingGroupAuthController.java
@@ -23,8 +23,8 @@ public class MeetingGroupAuthController {
 
     private final MeetingGroupService meetingGroupService;
     private static final String GET_GROUP = "/meeting-group";
-    private static final String GET_GROUP_CATEGORY = "/meeting-group/groupCategory";
-    private static final String GET_GROUP_TYPE = "/meeting-group/groupType";
+    private static final String GET_GROUP_CATEGORY = "/meeting-group/group-category";
+    private static final String GET_GROUP_TYPE = "/meeting-group/group-type";
     private static final String GET_GROUP_HOME = "/meeting-group/group";
     private static final String GET_MEETING_GROUP_ID = "/meeting-group/{groupId}";
